pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tentacle"
version = "0.0.1"
description = "TUI replacement for GitKraken!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=6.1.0",
    "GitPython>=3.1.42",
]

[project.scripts]
tentacle = "tentacle.main:main"

[tool.hatch.build.targets.wheel]
packages = ["tentacle"]
